+++ apps/hadron_prime/views/system/delete_hadron_system_views.py
@@ -21,7 +21,9 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import get_object_or_404, redirect
 from django.views.generic import TemplateView
 
+from apps.core.user_permissions.permission_manager import UserPermissionManager
 from apps.hadron_prime.models import HadronSystem
+from apps.user_permissions.utils import PermissionNames
 from web_project import TemplateLayout
 
 logger = logging.getLogger(__name__)
@@ -37,6 +39,15 @@ class HadronPrimeView_DeleteHadronSystem(LoginRequiredMixin, TemplateView):
 
     def post(self, request, *args, **kwargs):
         system_id = kwargs.get('pk')
+
+        ##############################
+        # PERMISSION CHECK FOR - DELETE_HADRON_SYSTEMS
+        if not UserPermissionManager.is_authorized(user=self.request.user,
+                                                   operation=PermissionNames.DELETE_HADRON_SYSTEMS):
+            messages.error(self.request, "You do not have permission to delete Hadron Systems.")
+            return redirect('hadron_prime:list_hadron_system')
+        ##############################
+
         hadron_system = get_object_or_404(HadronSystem, id=system_id)
         hadron_system.delete()
         logger.info(f'Hadron System "{hadron_system.system_name}" has been deleted successfully.')
