+++ apps/hadron_prime/views/topic/delete_hadron_topics_views.py
@@ -23,7 +23,9 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import get_object_or_404, redirect
 from django.views.generic import TemplateView
 
+from apps.core.user_permissions.permission_manager import UserPermissionManager
 from apps.hadron_prime.models import HadronTopic
+from apps.user_permissions.utils import PermissionNames
 from web_project import TemplateLayout
 
 logger = logging.getLogger(__name__)
@@ -39,6 +41,15 @@ class HadronPrimeView_DeleteHadronTopic(LoginRequiredMixin, TemplateView):
 
     def post(self, request, *args, **kwargs):
         topic_id = kwargs.get('pk')
+
+        ##############################
+        # PERMISSION CHECK FOR - DELETE_HADRON_TOPICS
+        if not UserPermissionManager.is_authorized(user=self.request.user,
+                                                   operation=PermissionNames.DELETE_HADRON_TOPICS):
+            messages.error(self.request, "You do not have permission to delete Hadron Topics.")
+            return redirect('hadron_prime:list_hadron_system')
+        ##############################
+
         hadron_topic = get_object_or_404(HadronTopic, id=topic_id)
         hadron_topic.delete()
         logger.info(f'Hadron Topic "{hadron_topic.topic_name}" has been deleted successfully.')
